test: enable MultiRangeInlineEditor-test
